refactor launch_ray and notebook init code
