+++ examples/ray/notebook/ray_determined/__init__.py
@@ -1,1 +1,1 @@
-from ray_determined.launch_ray import *
+from ray_determined._ray import init_ray_determined
