+++ examples/ray/notebook/ray_determined/startup-hook.sh
@@ -1,1 +1,2 @@
+# TODO: Remove after updating Docker image to install pyarrow.
 pip install pyarrow
